move testcase generation to utils
